clean up readme (#29)
* clean up readme

This change moves some of the more crunchy details of working with
oshinko-rest to the hacking document. It also cleans up the readme with
just some basic information about building and running the application.

* change contributing header style

This change makes the headers elements in the contributing document
markdown to be the same style as the other markdown documents.

* add an example title to run section

This just adds a bolded "Example" title to the section on running the
server, mainly to maintain consistency with the previous section.