enable tls access and upgrade generated code (#81)
* enable tls access and upgrade generated code

This change enables https access, to do this the generated code has been
updated as has the api definition. This change also required the
upgrading of the graceful lib in the vendor directory.

Changes
* add https scheme to openapi definition
* upgrade generated code, using go-swagger version 0.5.0-180-g1693567
* upgrade graceful vendored library to current release
* fix client tests to use http

* add scheme option to start script

This change will ensure that the server will only expose an http route.
This is being added to ensure backward compatiblity as we get closer to
a TLS solution.

* update readme with information about TLS

This adds a little documentation about how to invoke the server with
http and/or https support. It also adjusts the default invocation method
to only use http.

* fix small grammar mistake

Adding a missing word.

* fix spelling error

A word was spelled wrong, it has been "fixed".