Initial implementation for named cluster configs (#83)
* Initial implementation for named cluster configs

This change adds a "config" object to the NewCluster model which
now contains the master and worker counts along with a name field.
If the name is set, oshinko will look for the named configuration at
/etc/oshinko-cluster-configs and use that as the configuration for
the cluster with master and worker counts applied if they are
non-zero. For full details, see the cluster-configs.md spec
in the oshinko-specs repo.

* Fix existing unit tests to use config object

* Config object should contain configured master/worker count

The JSON responses from create, update, and find operations
should contain the configured master and worker node count, not
the actualy master and worker node count. The current master
and worker counts can be figured out from the pod list if
it's important. This is particularly important on create and
update where the response from the operation should inform
the caller of how many pods there will be (not how many
currently exist).

* Add additional error messages

Check for deployment configs on update (use this as our canonical
check for whether a cluster exists), add error handling on
failure to find replication controllers on update or list.

* Add unit tests for clusterconfigs helper

* Add the oshino-cluster-configs configmap to the template

This will create a base configmap for cluster configs and
mount it on the oshinko pod as a volume. The configmap
must exist for the volume mount to succeed. Include a
single definition for "small.workercount" as an example.

* Add a doc explaining named configurations and how to use them

* Tweak the text of the clusterconfigs doc

* Fix the api/swagger.yaml and update ClusterConfigs.md

Looks like the swagger definition should indeed include both
http and https as schemes. Also, make the curl example for
cluster create with a named config a little more explicit.

* Show in Clusterconfigs.md how to get a list of defined configurations