Allow master and worker spark configurations in cluster config
This change allows the name of a configmap to be specified for
the master and/or worker nodes in a spark cluster.  The configmap
will be mounted at a well-known location and the SPARK_CONF_DIR
env var will be set to that location. This lets users specify
spark configuration files for the nodes.