Add a build for an extended utility client
The spark application launch scripts need some additional CLI
commands to build and configure a cluster/dirver, but don't need
to include the entire "oc" client. This change adds a build for an
extended oshinko CLI called "oshinko-clix" which is a
CLI with a superset of the oshinko-cli commands. The implementation
of the oshinko-cli subset is shared.