Prototype metrics and scorpionstare implementation
This change supports enabling a graphite metrics server
with a carbon store on the spark master.  It also supports
adding a scorpionstare container on the spark master for
dynamic scaling. Note that at this point it's necessary
to supply the metrics.properties configuration on the
spark master manually (either in the image or via
a configmap at launch time)