Pass --image flag value in cluster config object
This change brings oshinko-cli up to date with changes
in oshinko-core to allow the spark image to be specified in
the cluster config. Oshinko-core now treats the "image"
parameter to create() as a default if a cluster config does not
specify an image value. Consequently, the CLI should pass the
default spark image as the "image" parameter and pass any
value from the '--image' flag in a config object.  This
will result in the --image flag value having the highest precedence,
followed by an image value set in a stored config, followed by the
default spark image set in the CLI.

Updates oshinko-core in the vendor dir:

    Allow spark image to be specified in clusterconfigs

    The create routine already allows the image value to be passed
    in as an argument. Allow the image to be set in a cluster config
    as well, which will take precedence. This

    (commit c3af0d476a5c6592fd54847dd013cd3000cf6745 in oshinko-core)