Reconcile file set with oshinko-cli and oshinko-rest master
The merge of oshinko-rest into ./rest to preserve history had some
conflicts in top level files like README.md and Makefile between
the two repos. Also, the vendor dirs between oshinko-cli and
oshinko-rest had to be merged.

This commit reconciles the file set with the master branch of
oshinko-cli and oshinko-rest so that the diffs are clean.