Allow creation of clusters with zero masters
This change is related to recent changes which allow scaling
masters and workers to zero. Creation needs consistency, and
bugs prevented a cluster with master count of zero. Note that
in a cluster created without a master, the workers will loop
forever waiting for the master to be available (ie if/when the
master count is scaled to 1)