Support returning incomplete clusters with Incomplete status
This change reports clusters which are missing components and
gives their status as Incomplete, along with some sentinel values
for certain fields. This is useful in cases where clusters are in
the process of being created or deleted, since these operations from
the kube/openshift perspective are not atomic.

Note, refactored FindSingleCluster and FindClusters to use the
same base routine for reporting cluster components.