Retry update on driver rc during create
Kubernetes will give an error on an update if the object
being updated has changed between the time it was retrieved and
then update call. So, when we add the driver label to the
app deployment in an ephemeral cluster case, retry the update
for up to 5 seconds before giving up.