Don't consider pod list when determining Incomplete
Oshinko currently cannot tell if pods are Running or Terminating.
If a cluster is otherwise Incomplete (ie there are no services
or deploymentconfigs) do not look at the pod list when deciding
between Incomplete and "no cluster".  If the dcs are gone, the
pods must be terminating, and conflicts will not arise if a
new cluster of the same name is created.