Move the --app flag for get into extended
All of the ephemeral cluster options are in the extended
cli, available with "make extended". Put the --app flag
for get there, too. It's already apparent from normal get
which clusters are shared and which belong to certain apps,
you can find it with grep.

We'll have to decide in the future whether to put the ephemeral
stuff in the regular CLI. At this point, it's really only useful
in CLI workflows.  Another application that rides on top of oshinko
can track it's own state and manage "ephemeral-ness" by itself,
or just use an S2I workflow.