Look for dcs rather than pods in FindClusters
The "list all" method for clusters was looking at oshinko
pods to find cluster names, but if we allow scale to zero
of both workers and masters then there are no pods from
which to discover names, of course.  Use deploymentconfigs
instead. At present, if a master deploymentconfig doesn't
exist for a cluster then the cluster is assumed not to exist.