update vendor to latest 1.4 origin release
This change updates the vendor directory to use the latest OpenShift
Origin sources for the 1.4 release line. It also adds the Godeps.json
file to the top level Godeps directory. This change has been added to
help improve the quality of the dependency chain for the oshinko cli and
rest applications.

In addition to the vendor changes, the Godeps.json file has been
modified by hand to include the dependencies for the rest server. These
changes have been made by hand to account for difficulties with the
godep tooling when evaluating the OpenShift code base. As OpenShift
Origin uses modified upstreams in its vendor directory, these have been
included in the oshinko vendor.

The CONTRIBUTING.md file has also been updated to record the process
used to create the vendor directory and the Godeps.json file.