Add additional tests
This change adds tests for the hidden commands. It also expands
the coverage for the existing commands and splits up tests
into multiple files.

Additionally, it adds a test runner "test/run.sh" which can
be used to run the same tests against an OpenShift instance
instead of running them with hack/test-cmd.sh.