Set spark default image for oshinko-rest in build.
This change creates a common sparkimage.sh file to hold
a default spark image spec that is used at build time for
both oshinko-rest and oshinko cli.

For oshinko-rest, it's still possible to override the default
with an environment variable, but if the env var is left unspecified
then the build-time setting will be used.

It also adds the default image value to the server response.