Improvement of REST deletion
Since deletion of a spark cluster is not atomic (services,
dcs, rcs, etc) oshinko reports extended information if
some items are not deleted cleanly, but does not return
an error if at least some part of the cluster was
deleted.  The CLI displays this info and does not
return an error, the REST interface should do the
same.

Additionaly, make the client test report deletion error
detail.